long's not ints for the save call

+++ web_module/trunk/src/main/com/topcoder/web/codinginterface/longcontest/controller/request/Submit.java
@@ -372,7 +372,7 @@ public class Submit extends Base {
         TestServices ts = t.create();
 
         // Save the code!
-        return ts.saveLongComponent((int)cd, (int)rid, (int)cid, (int)uid, code, lang).isSuccess();
+        return ts.saveLongComponent(cd, rid, cid, uid, code, lang).isSuccess();
 
     }
